mtd-utils: fix source revision
The PKG_REV was improperly defined causing mtd-utils master to be used instead
of the v1.5.2 tag. This caused the mtd-utils package to be unbuildable after
upstream moved to autotools for v2.0.0.

Add PKG_REV into the PKG_SOURCE as well to force a re-fetch of the code.
Users should remove inappropriate dl/mtd-utils-1.5.2.tar.gz to avoid confusion.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>