add gw16159 driver and firmware packages
The GW16159 is an M.2 E-Key card with a Silex SX-SDMAH module consisting
of a MorseMicro SDIO based mm6108 802.11ah (HaLow) radio.

The driver and firmware packages (morse_driver and morse-fw) include the
driver from https://github.com/gateworks/morse_driver (which is the
MorseMicro upstream driver with some patches for GW16159 compatiblity)
and the required firmware.

OpenWrt will see the radio as a generic MAC80211 802.11ac/n radio and
allowed it to be configured as an AP/STA.

For more details see: https://trac.gateworks.com/wiki/expansion/gw16159

Signed-off-by: Tim Harvey <tharvey@gateworks.com>